+++ vscode/e2e/default_settings/backlink.test.ts
@@ -15,8 +15,17 @@ describe('backlinks', () => {
     noteSearcher = new NoteSearcherDriver(vscode);
   });
 
-  it('clicking on a backlink goes to that note', async () => {
+  it('clicking on a backlink opens the backlink', async () => {
     await vscode.openDemoDirFile('trains.md');
+    const backlink = await noteSearcher.findBacklinkByName('readme.md');
+    expect(backlink).not.to.be.null;
+    await backlink!.click();
+    const currentEditor = await vscode.currentEditor();
+    expect(await currentEditor!.getTitle()).to.equal('readme.md');
+  });
+
+  it('backlinks support markdown links', async () => {
+    await vscode.openDemoDirFile('cheese.md');
     const backlink = await noteSearcher.findBacklinkByName('readme.md');
     expect(backlink).not.to.be.null;
     await backlink!.click();