+++ vscode/src/index/lunrNoteIndex.ts
@@ -23,11 +23,11 @@ export class LunrNoteIndex implements NoteIndex {
 
   public containsNote = (path: string) => this._linkIndex.containsNote(path);
 
-  public markdownLinksFrom = (path: string) => this._linkIndex.markdownLinksFrom(path);
-
-  public wikiLinksFrom = (path: string) => this._linkIndex.wikiLinksFrom(path);
+  public linksFrom = (path: string) => this._linkIndex.linksFrom(path);
 
   public linksTo = (path: string) => this._linkIndex.linksTo(path);
+
+  public findAllDeadLinks = () => this._linkIndex.findAllDeadLinks();
 
   private indexAllFiles = async (dir: string) => {
     this._tags.clear();
@@ -41,6 +41,7 @@ export class LunrNoteIndex implements NoteIndex {
     }
 
     await Promise.all(jobs);
+    this._linkIndex.finalise();
     this._lunrSearch.finalise();
   };