+++ vscode/src/index/mapLinkIndex.deadLinks.test.ts
@@ -2,23 +2,21 @@ const _path = require('path');
 
 import * as tmoq from 'typemoq';
 
-import { DeadLinkFinder } from "./DeadLinkFinder";
-import { MapLinkIndex } from "../index/noteLinkIndex";
+import { MapLinkIndex } from "./noteLinkIndex";
 import { MockFile } from "../mocks/MockFile";
-import { LunrNoteIndex } from "../index/lunrNoteIndex";
+import { LunrNoteIndex } from "./lunrNoteIndex";
 import { createFileSystem, FileSystem } from "../utils/FileSystem";
-import { NoteIndex } from "../index/NoteIndex";
 
-describe('dead link finder, mocked filesystem', () => {
+describe('MapLinkIndex, dead links, mocked filesystem', () => {
   let fileSystem: tmoq.IMock<FileSystem>;
   let linkIndex: MapLinkIndex;
-  let finder: DeadLinkFinder;
 
   const setupLinks = (fileLinks: MockFile[]) => {
     for (const file of fileLinks) {
       linkIndex.addFile(file.path(), file.text());
       fileSystem.setup(f => f.fileExists(file.path())).returns(() => true);
     }
+    linkIndex.finalise();
   };
 
   describe('posix paths, markdown links', () => {
@@ -27,7 +25,6 @@ describe('dead link finder, mocked filesystem', () => {
     beforeEach(() => {
       fileSystem = tmoq.Mock.ofType<FileSystem>();
       linkIndex = new MapLinkIndex();
-      finder = new DeadLinkFinder(linkIndex, fileSystem.object);
     });
 
     it('finds dead link', () => {
@@ -35,7 +32,7 @@ describe('dead link finder, mocked filesystem', () => {
         new MockFile('/a.md', '[](/b.txt)')
       ]);
 
-      const deadLinks = finder.findAllDeadLinks();
+      const deadLinks = linkIndex.findAllDeadLinks();
 
       expect(deadLinks).toHaveLength(1);
       expect(deadLinks[0].sourcePath).toBe('/a.md');
@@ -48,7 +45,7 @@ describe('dead link finder, mocked filesystem', () => {
         new MockFile('/b.txt', '')
       ]);
 
-      const deadLinks = finder.findAllDeadLinks();
+      const deadLinks = linkIndex.findAllDeadLinks();
 
       expect(deadLinks).toHaveLength(0);
     });
@@ -59,7 +56,7 @@ describe('dead link finder, mocked filesystem', () => {
         new MockFile('/b/c/e.txt', '')
       ]);
 
-      const deadLinks = finder.findAllDeadLinks();
+      const deadLinks = linkIndex.findAllDeadLinks();
 
       expect(deadLinks).toHaveLength(0);
     });
@@ -70,7 +67,7 @@ describe('dead link finder, mocked filesystem', () => {
         new MockFile('/a.md', '')
       ]);
 
-      const deadLinks = finder.findAllDeadLinks();
+      const deadLinks = linkIndex.findAllDeadLinks();
 
       expect(deadLinks).toHaveLength(0);
     });
@@ -81,7 +78,7 @@ describe('dead link finder, mocked filesystem', () => {
         new MockFile('/a.md', '')
       ]);
 
-      const deadLinks = finder.findAllDeadLinks();
+      const deadLinks = linkIndex.findAllDeadLinks();
 
       expect(deadLinks).toHaveLength(0);
     });
@@ -92,7 +89,7 @@ describe('dead link finder, mocked filesystem', () => {
         new MockFile('/a/c/d.txt', '')
       ]);
 
-      const deadLinks = finder.findAllDeadLinks();
+      const deadLinks = linkIndex.findAllDeadLinks();
 
       expect(deadLinks).toHaveLength(0);
     });
@@ -103,7 +100,7 @@ describe('dead link finder, mocked filesystem', () => {
         new MockFile('/a/c.png', '')
       ]);
 
-      const deadLinks = finder.findAllDeadLinks();
+      const deadLinks = linkIndex.findAllDeadLinks();
 
       expect(deadLinks).toHaveLength(0);
     });
@@ -115,7 +112,6 @@ describe('dead link finder, mocked filesystem', () => {
     beforeEach(() => {
       fileSystem = tmoq.Mock.ofType<FileSystem>();
       linkIndex = new MapLinkIndex();
-      finder = new DeadLinkFinder(linkIndex, fileSystem.object);
     });
 
     it('finds dead link', () => {
@@ -123,7 +119,7 @@ describe('dead link finder, mocked filesystem', () => {
         new MockFile('/a.md', '[[blah]]')
       ]);
 
-      const deadLinks = finder.findAllDeadLinks();
+      const deadLinks = linkIndex.findAllDeadLinks();
 
       expect(deadLinks).toHaveLength(1);
       expect(deadLinks[0].sourcePath).toBe('/a.md');
@@ -136,7 +132,7 @@ describe('dead link finder, mocked filesystem', () => {
         new MockFile('/b.txt', '')
       ]);
 
-      const deadLinks = finder.findAllDeadLinks();
+      const deadLinks = linkIndex.findAllDeadLinks();
 
       expect(deadLinks).toHaveLength(0);
     });
@@ -147,7 +143,7 @@ describe('dead link finder, mocked filesystem', () => {
         new MockFile('/b/c/e.txt', '')
       ]);
 
-      const deadLinks = finder.findAllDeadLinks();
+      const deadLinks = linkIndex.findAllDeadLinks();
 
       expect(deadLinks).toHaveLength(0);
     });
@@ -158,7 +154,7 @@ describe('dead link finder, mocked filesystem', () => {
         new MockFile('/a.md', '')
       ]);
 
-      const deadLinks = finder.findAllDeadLinks();
+      const deadLinks = linkIndex.findAllDeadLinks();
 
       expect(deadLinks).toHaveLength(0);
     });
@@ -170,7 +166,6 @@ describe('dead link finder, mocked filesystem', () => {
     beforeEach(() => {
       fileSystem = tmoq.Mock.ofType<FileSystem>();
       linkIndex = new MapLinkIndex();
-      finder = new DeadLinkFinder(linkIndex, fileSystem.object);
     });
 
     it('finds dead link', () => {
@@ -178,7 +173,7 @@ describe('dead link finder, mocked filesystem', () => {
         new MockFile('c:\\a.md', '[](c:\\b.txt)')
       ]);
 
-      const deadLinks = finder.findAllDeadLinks();
+      const deadLinks = linkIndex.findAllDeadLinks();
 
       expect(deadLinks).toHaveLength(1);
       expect(deadLinks[0].sourcePath).toBe('c:\\a.md');
@@ -191,7 +186,7 @@ describe('dead link finder, mocked filesystem', () => {
         new MockFile('c:\\b.txt', '')
       ]);
 
-      const deadLinks = finder.findAllDeadLinks();
+      const deadLinks = linkIndex.findAllDeadLinks();
 
       expect(deadLinks).toHaveLength(0);
     });
@@ -202,7 +197,7 @@ describe('dead link finder, mocked filesystem', () => {
         new MockFile('c:\\b\\c\\e.txt', '')
       ]);
 
-      const deadLinks = finder.findAllDeadLinks();
+      const deadLinks = linkIndex.findAllDeadLinks();
 
       expect(deadLinks).toHaveLength(0);
     });
@@ -213,7 +208,7 @@ describe('dead link finder, mocked filesystem', () => {
         new MockFile('c:\\a.md', '')
       ]);
 
-      const deadLinks = finder.findAllDeadLinks();
+      const deadLinks = linkIndex.findAllDeadLinks();
 
       expect(deadLinks).toHaveLength(0);
     });
@@ -224,7 +219,7 @@ describe('dead link finder, mocked filesystem', () => {
         new MockFile('c:\\a.md', '')
       ]);
 
-      const deadLinks = finder.findAllDeadLinks();
+      const deadLinks = linkIndex.findAllDeadLinks();
 
       expect(deadLinks).toHaveLength(0);
     });
@@ -235,7 +230,7 @@ describe('dead link finder, mocked filesystem', () => {
         new MockFile('c:\\a\\c\\d.txt', '')
       ]);
 
-      const deadLinks = finder.findAllDeadLinks();
+      const deadLinks = linkIndex.findAllDeadLinks();
 
       expect(deadLinks).toHaveLength(0);
     });
@@ -246,7 +241,7 @@ describe('dead link finder, mocked filesystem', () => {
         new MockFile('c:\\a\\c.png', '')
       ]);
 
-      const deadLinks = finder.findAllDeadLinks();
+      const deadLinks = linkIndex.findAllDeadLinks();
 
       expect(deadLinks).toHaveLength(0);
     });
@@ -258,7 +253,6 @@ describe('dead link finder, mocked filesystem', () => {
     beforeEach(() => {
       fileSystem = tmoq.Mock.ofType<FileSystem>();
       linkIndex = new MapLinkIndex();
-      finder = new DeadLinkFinder(linkIndex, fileSystem.object);
     });
 
     it('finds dead link', () => {
@@ -266,7 +260,7 @@ describe('dead link finder, mocked filesystem', () => {
         new MockFile('c:\\a.md', '[[blah]]')
       ]);
 
-      const deadLinks = finder.findAllDeadLinks();
+      const deadLinks = linkIndex.findAllDeadLinks();
 
       expect(deadLinks).toHaveLength(1);
       expect(deadLinks[0].sourcePath).toBe('c:\\a.md');
@@ -279,7 +273,7 @@ describe('dead link finder, mocked filesystem', () => {
         new MockFile('c:\\b.txt', '')
       ]);
 
-      const deadLinks = finder.findAllDeadLinks();
+      const deadLinks = linkIndex.findAllDeadLinks();
 
       expect(deadLinks).toHaveLength(0);
     });
@@ -290,7 +284,7 @@ describe('dead link finder, mocked filesystem', () => {
         new MockFile('c:\\b\\c\\e.txt', '')
       ]);
 
-      const deadLinks = finder.findAllDeadLinks();
+      const deadLinks = linkIndex.findAllDeadLinks();
 
       expect(deadLinks).toHaveLength(0);
     });
@@ -301,50 +295,33 @@ describe('dead link finder, mocked filesystem', () => {
         new MockFile('c:\\a.md', '')
       ]);
 
-      const deadLinks = finder.findAllDeadLinks();
+      const deadLinks = linkIndex.findAllDeadLinks();
 
       expect(deadLinks).toHaveLength(0);
     });
   });
 });
 
-describe('dead link finder, real filesystem', () => {
-  let linkIndex: NoteIndex;
-  let finder: DeadLinkFinder;
+describe('MapLinkIndex, dead links, real filesystem', () => {
+  let linkIndex: LunrNoteIndex;
 
   beforeEach(() => {
     const fs = createFileSystem();
     linkIndex = new LunrNoteIndex(fs);
-    finder = new DeadLinkFinder(linkIndex, fs);
   });
 
-  it('finds dead markdown link in demo dir readme', async () => {
+  it('finds all dead links in demo dir', async () => {
     await linkIndex.index(_path.resolve(__dirname, '../../demo_dir'));
 
     // act
-    const deadLinks = finder.findAllDeadLinks();
+    const deadLinks = linkIndex.findAllDeadLinks();
 
     // assert
     expect(deadLinks).toHaveLength(2);
+    expect(deadLinks.map(d => _path.parse(d.sourcePath).base))
+      .toStrictEqual(['readme.md', 'readme.md']);
 
-    const deadMarkdownLink = deadLinks
-      .filter(d => d.targetPath.includes('to/nowhere'))[0];
-
-    expect(deadMarkdownLink.sourcePath).toContain('readme.md');
-  });
-
-  it('finds dead wiki link in demo dir readme', async () => {
-    await linkIndex.index(_path.resolve(__dirname, '../../demo_dir'));
-
-    // act
-    const deadLinks = finder.findAllDeadLinks();
-
-    // assert
-    expect(deadLinks).toHaveLength(2);
-
-    const deadMarkdownLink = deadLinks
-      .filter(d => d.targetPath.includes('non_existent_note'))[0];
-
-    expect(deadMarkdownLink.sourcePath).toContain('readme.md');
+    expect(deadLinks.map(d => _path.parse(d.targetPath).base))
+      .toStrictEqual(['nowhere.md', 'non_existent_note']);
   });
 });