+++ vscode/CHANGELOG.md
@@ -3,6 +3,13 @@
 All notable changes to the "note-searcher" extension will be documented in this
 file. Format inspired by https://keepachangelog.com/en/1.0.0/
 
+## [0.0.24] - 2020-12-28
+### Added
+- backlinks now support wiki links
+
+### Changed
+- wiki links within code blocks are ignored
+
 ## [0.0.23] - 2020-12-06
 ### Added
 - copy wiki link from file explorer